+++ Cryptodd/Binance/Orderbooks/Handlers/BinanceOrderbookHandlerArguments.cs
@@ -1,4 +1,4 @@
-﻿namespace Cryptodd.Binance.Orderbook.Handlers;
+﻿namespace Cryptodd.Binance.Orderbooks.Handlers;
 
 public record BinanceOrderbookHandlerArguments(
     string Symbol,
