+++ Cryptodd/Binance/Orderbooks/InMemoryOrderbook.SortedView.cs
@@ -3,7 +3,7 @@ using System.Collections.Concurrent;
 using System.Runtime.CompilerServices;
 using Maxisoft.Utils.Collections.Lists.Specialized;
 
-namespace Cryptodd.Binance.Orderbook;
+namespace Cryptodd.Binance.Orderbooks;
 
 public partial class InMemoryOrderbook<T>
 {
