+++ Cryptodd/Binance/Orderbooks/Websocket/BinanceOrderbookWebsocketOptions.cs
@@ -1,3 +1,3 @@
-﻿namespace Cryptodd.Binance.Orderbook.Websocket;
+﻿namespace Cryptodd.Binance.Orderbooks.Websocket;
 
 public class BinanceOrderbookWebsocketOptions : BaseBinanceOrderbookWebsocketOptions { }
