+++ Cryptodd/Binance/Orderbooks/Websocket/BinanceWebsocketStats.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Concurrent;
 
-namespace Cryptodd.Binance.Orderbook.Websocket;
+namespace Cryptodd.Binance.Orderbooks.Websocket;
 
 public readonly record struct PerSymbolStatsEntry(string Symbol, long CallCounter, DateTimeOffset LastCall) { }
